// src/shared/config/enums.ts
export enum ItemType {
  EGG = 'egg',
  PET = 'pet',
  POTION = 'potion',
  TRANSPORT = 'transport',
}

export enum Rarity {
  COMMON = 'common',
  UNCOMMON = 'uncommon',
  RARE = 'rare',
  ULTRA_RARE = 'ultra_rare',
  LEGENDARY = 'legendary',
}

export enum Property {
  DEFAULT = 'default',
  NEON = 'neon',
  MEGA_NEON = 'mega_neon',
  FLYABLE = 'flyable',
  RIDEABLE = 'rideable',
}

export enum SortType {
  POPULARITY = 'popularity',
  PRICE = 'price',
}

export enum SortDirection {
  ASC = 'asc',
  DESC = 'desc',
}